Move Hostname to its own files.
Also, add resolve-related variables and functions in Hostname.
